% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-run.R
\name{activate_plate_run}
\alias{activate_plate_run}
\title{Activate Plate Run}
\usage{
activate_plate_run(con, plate_run_id)
}
\arguments{
\item{con}{valid connection to the database}

\item{plate_run_id}{plate run identifier}
}
\value{
no return value
}
\description{
\code{activate_plate_run()} activates an existing plate run entry in
the plate_run table in the database.
}
