% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{add_assay_results}
\alias{add_assay_results}
\title{Add assay results to run-id-database}
\usage{
add_assay_results(con, assay_results)

add_assay_results(con, assay_results)
}
\description{
Add assay results to run-id-database

Add Raw Results
}
