% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-genetic-method.R
\name{add_genetic_method}
\alias{add_genetic_method}
\title{Add Genetic Method}
\usage{
add_genetic_method(con, genetic_method)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{genetic_method}{A valid genetic method dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} 4 letter code in all caps
\item \strong{method_name} \emph{character} Short name for method
\item \strong{description} \emph{character} Brief description of method
}}
}
\description{
\code{add_genetic_method()} adds a new genetic method type to the genetic method lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)


new_method <- data.frame(code = "FAKE", method_name = "Fake Method",
                         description = "A fake method created for this example")
add_genetic_method(con, new_method)
}
\seealso{
Other genetic method functions:
\code{\link[=delete_genetic_method]{delete_genetic_method()}},
\code{\link[=get_genetic_methods]{get_genetic_methods()}},
\code{\link[=update_genetic_method]{update_genetic_method()}},
\code{\link[=update_genetic_method_status]{update_genetic_method_status()}}
}
\concept{genetic method functions}
