% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-laboratory.R
\name{add_laboratory}
\alias{add_laboratory}
\title{Add Laboratory}
\usage{
add_laboratory(con, laboratory)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{laboratory}{A valid laboratory dataframe containing the following:
\itemize{
\item \strong{code} Short code representing laboratory name
\item \strong{laboratory_name} Laboratory full name
\item \strong{description} Brief description of the laboratory
}}
}
\description{
\code{add_laboratory()} adds a new laboratory to the laboratory lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_laboratory <- data.frame(code = "CNL",
                             laboratory_name = "Cool New Lab",
                             description = "The coolest new lab ever")
add_laboratory(con, new_laboratory)
}
\seealso{
Other laboratory functions:
\code{\link[=delete_laboratory]{delete_laboratory()}},
\code{\link[=get_laboratories]{get_laboratories()}},
\code{\link[=update_laboratory]{update_laboratory()}},
\code{\link[=update_laboratory_status]{update_laboratory_status()}}
}
\concept{laboratory functions}
