% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.R
\name{add_personnel}
\alias{add_personnel}
\title{Add personnel to table}
\usage{
add_personnel(con, first_name, last_name, laboratory_id)
}
\arguments{
\item{con}{connection to database}

\item{first_name}{first name for personnel}

\item{last_name}{last name for personnel}

\item{laboratory_id}{the id for lab personnel is associated with}
}
\description{
Add personnel to table
}
