% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{add_plate_run}
\alias{add_plate_run}
\title{Create Plate Run}
\usage{
add_plate_run(
  con,
  protocol_id,
  genetic_method_id,
  laboratory_id,
  lab_work_preformed_by,
  date_run
)
}
\description{
blah
}
