% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{add_plate_thresholds}
\alias{add_plate_thresholds}
\title{Add Thresholds to Database}
\usage{
add_plate_thresholds(
  con,
  thresholds,
  destination_table,
  results_table,
  .control_id = "EBK"
)
}
\arguments{
\item{con}{a connection the database}

\item{thresholds}{a thresholds table to use as insert data, this table is generated by \code{generate_threshold}}

\item{.control_id}{the id to use as the control for thresholds}
}
\description{
Add Thresholds to Database
}
\details{
Additional details...
}
