% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-protocol.R
\name{add_protocol}
\alias{add_protocol}
\title{Add Protocol}
\usage{
add_protocol(con, protocol)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{protocol}{A valid protocol dataframe, reference \code{\link{protocol_template}} as an example}
}
\description{
\code{add_protocol()} adds a new protocol to the protocol lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_protocol <- protocol_template
new_protocol$software_version <- "3.11.20"
add_protocol(con, new_protocol)
}
\seealso{
Other protocol functions:
\code{\link[=add_protocol_based_on]{add_protocol_based_on()}},
\code{\link[=delete_protocol]{delete_protocol()}},
\code{\link[=get_protocols]{get_protocols()}},
\code{\link[=update_protocol]{update_protocol()}},
\code{\link[=update_protocol_status]{update_protocol_status()}}
}
\concept{protocol functions}
