% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{add_raw_assay_results}
\alias{add_raw_assay_results}
\title{Add Raw Results}
\usage{
add_raw_assay_results(
  con,
  assay_results,
  destination_table = c("raw_assay_result", "externtal_raw_assay_result")
)
}
\description{
Add Raw Results
}
