% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-run-type.R
\name{add_run_type}
\alias{add_run_type}
\title{Add Run Type}
\usage{
add_run_type(con, run_type)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{run}{type A valid run type dataframe containing the following:
\itemize{
\item \strong{code} Short code representing run type name
\item \strong{run_name} Run type full name
\item \strong{description} Brief description of the run type
}}
}
\description{
\code{add_run_type()} adds a new run type to the run type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_run_type <- data.frame(code = "SWR",
                           run_name = "Spring Winter Run",
                           description = "A pretend run")
add_run_type(con, new_run_type)
}
\seealso{
Other run type functions:
\code{\link[=delete_run_type]{delete_run_type()}},
\code{\link[=get_run_types]{get_run_types()}},
\code{\link[=update_run_type]{update_run_type()}},
\code{\link[=update_run_type_status]{update_run_type_status()}}
}
\concept{run type functions}
