% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvage-plate-run.R
\name{add_salvage_samples}
\alias{add_salvage_samples}
\title{Seed Salvate Samples}
\usage{
add_salvage_samples(con, n_samples = 1000, year)
}
\arguments{
\item{con}{a conection the database}

\item{n_samples}{total number of samples to create}
}
\description{
Given a total number of samples, this fucntion creates and inserts them into the "external_sample" table
in the database connected via the \code{con} object.
}
