% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-events.R
\name{add_sample}
\alias{add_sample}
\title{Add single or many samples}
\usage{
add_sample(
  con,
  location_code,
  sample_event_number,
  first_sample_date,
  sample_bin_code,
  min_fork_length,
  max_fork_length,
  expected_number_of_samples
)
}
\description{
The intended use for this function to add samples when additional samples are received from field activity.
}
