% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{add_sherlock_results}
\alias{add_sherlock_results}
\title{Insert all results from Sherlock Output}
\usage{
add_sherlock_results(con, transformed_assay_results, sample_details)
}
\arguments{
\item{con}{a connection to the database}

\item{transformed_assay_results}{a list of containing two tables, \code{raw_assay_results} and \code{assay_results} result from \code{process_sherlock}}

\item{sample_details}{dataframe containing sample information at each well}
}
\description{
Adds results from Sherlock output after being processed via \code{process_sherlock()}
to the run-id database, as well as computed thresholds for run assignment.
}
\examples{
sample_details = readr::read_csv("data-raw/sample_layout_template.csv")
processed_results <- process_sherlock(filepath = "data-raw/exampleoutput_synergyH1trial_data_092021.xlsx",
                 sample_details = sample_details,
                 plate_size = 96)

add_sherlock_results(con, processed_results, sample_details)
}
