% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-status-code.R
\name{add_status_code}
\alias{add_status_code}
\title{Add Status Code}
\usage{
add_status_code(con, status_code)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{status_code}{A valid status code dataframe containing the following:
\itemize{
\item \strong{status_code} Status code full name
\item \strong{description} Brief description of the status code
}}
}
\description{
\code{add_status_code()} adds a new status code to the status code lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_status_code <- data.frame(status_code = "All done",
                              description = "Everything is all done")
add_status_code(con, new_status_code)
}
\seealso{
Other status code functions:
\code{\link[=delete_status_code]{delete_status_code()}},
\code{\link[=get_sample_status]{get_sample_status()}},
\code{\link[=get_status_codes]{get_status_codes()}},
\code{\link[=set_sample_status]{set_sample_status()}},
\code{\link[=update_status_code]{update_status_code()}},
\code{\link[=update_status_code_status]{update_status_code_status()}}
}
\concept{status code functions}
