% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add_to_config}
\alias{add_to_config}
\title{Set values in config file}
\usage{
add_to_config(key, value, file = get_config_file())
}
\arguments{
\item{key}{the key to update}

\item{value}{the value to update the key to}

\item{file}{config file, will use the file that config package finds by default}
}
\description{
Set values in config file
}
