% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_codes.R
\name{assign_run_types}
\alias{assign_run_types}
\title{assign run types}
\usage{
assign_run_types(assay_detections_with_status_codes)
}
\arguments{
\item{assay_detections}{table of assay detections}
}
\value{
a table with `run_type_id` column
}
\description{
helper function called within `add_genetic_identification`
}
\details{
`assign_run_types` takes in assay detections and assigns a run type ID
based on assay results logic and status codes.
}
\keyword{internal}
