% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.R
\name{delete_personnel}
\alias{delete_personnel}
\title{Remove personnel}
\usage{
delete_personnel(con, id)
}
\arguments{
\item{con}{connection to the database}

\item{id}{id for record to remove}
}
\description{
Remove personnel
}
\details{
Records are not actually deleted via this API, instead they are just set to inactive.
}
