% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-run.R
\name{delete_plate_run}
\alias{delete_plate_run}
\title{Delete Plate Run}
\usage{
delete_plate_run(con, plate_run_id)
}
\arguments{
\item{con}{valid connection to the database}

\item{plate_run_id}{plate run identifier}
}
\value{
no return value
}
\description{
`delete_plate_run()` deactivates an existing plate run entry in
the plate_run table in the database.
}
