% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-run-type.R
\name{delete_run_type}
\alias{delete_run_type}
\title{Delete Run Type}
\usage{
delete_run_type(con, run_type_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{run_type_id}{A numeric ID for the targeted run type \code{\link{get_run_types}}
\strong{Note:} If an run type has been associated with a sample, then
the database restricts deleting this run type. You must first update those records
with a new run type before reattempting to delete the run type. Consider
using the \code{\link{update_run_type_status}} function if you are wanting to
retire an run type while retaining its value for historic records.}
}
\description{
\code{delete_run_type()} deletes an existing run type in the run type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_run_types <- get_run_types(con)
View(all_run_types) # to view the ID of the run type needing deletion

delete_run_type(con, 1)
}
\seealso{
Other run type functions:
\code{\link[=add_run_type]{add_run_type()}},
\code{\link[=get_run_types]{get_run_types()}},
\code{\link[=update_run_type]{update_run_type()}},
\code{\link[=update_run_type_status]{update_run_type_status()}}
}
\concept{run type functions}
