% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-status-code.R
\name{delete_status_code}
\alias{delete_status_code}
\title{Delete Status Code}
\usage{
delete_status_code(con, status_code_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{status_code_id}{A numeric ID for the targeted status code \code{\link{get_status_codes}}
\strong{Note:} If a status code has been associated with a sample, then
the database restricts deleting this status code. You must first update those records
with a new status code before reattempting to delete the status code. Consider
using the \code{\link{update_status_code_status}} function if you are wanting to
retire an status code while retaining its value for historic records.}
}
\description{
\code{delete_status_code()} deletes an existing status code in the status code lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_status_codes <- get_status_codes(con)
View(all_status_codes) # to view the ID of the status code needing deletion

delete_status_code(con, 1)
}
\seealso{
Other status code functions:
\code{\link[=add_status_code]{add_status_code()}},
\code{\link[=get_sample_status]{get_sample_status()}},
\code{\link[=get_status_codes]{get_status_codes()}},
\code{\link[=set_sample_status]{set_sample_status()}},
\code{\link[=update_status_code]{update_status_code()}},
\code{\link[=update_status_code_status]{update_status_code_status()}}
}
\concept{status code functions}
