% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excel_column_index}
\alias{excel_column_index}
\title{Excel Column Index}
\format{
An object of class \code{character} of length 702.
}
\usage{
excel_column_index
}
\description{
Letter indices used to reference excel columns
}
\keyword{datasets}
