% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{extract_previous_end_row}
\alias{extract_previous_end_row}
\title{Extract Previous End Row}
\usage{
extract_previous_end_row(cell_ranges)
}
\arguments{
\item{cell_ranges}{a vector of strings representing the cell ranges of the tables e.g. c("A1:B10", "A11:B21")}
}
\description{
helper function used to identify the previous ending range of cells
within the Synergy H1 excel workbook output
}
