% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_sherlock_protocol}
\alias{extract_sherlock_protocol}
\title{Extract Sherlock Protocol}
\usage{
extract_sherlock_protocol(filepath)
}
\description{
helper function for parsing protocol settings within SynergyH1 output
}
\details{
TODO
}
