% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_sherquant_protocol}
\alias{extract_sherquant_protocol}
\title{Extract SherQUANT Protocol}
\usage{
extract_sherquant_protocol(filepath)
}
\description{
helper function for parsing protocol settings within SherQuant output
}
\details{
This assumes that if 41 cycles are performed, the protocol was a 2 hour run time.
}
