% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{fill_dual_assay_plate_map}
\alias{fill_dual_assay_plate_map}
\title{fill_dual_assay_plate_map}
\usage{
fill_dual_assay_plate_map(sample_ids)
}
\arguments{
\item{sample_ids}{list of sample IDs with which to populate a plate map}
}
\value{
A list of plate map tables
}
\description{
Generates a dual assay 384-well plate map from a list of sample IDs
}
\details{
Generate Plate Map for Dual Assay layout

This is called in \code{generate_subsample_plate_map()} and fills in alternating rows first, and then
the rows inbetween on a 384-well plate. Sample IDs are populated in the left half of a 384-well plate
and then replicated on the right half of the plate with control blanks in rows 12 and 24.
}
