% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_assay_detection}
\alias{generate_assay_detection}
\title{Generate detection for assay results offline}
\usage{
generate_assay_detection(
  thresholds,
  offline_sherlock_results,
  .control_id = "NTC"
)
}
\arguments{
\item{thresholds}{threshold values calculated in `generate_threshold_offline`}

\item{offline_sherlock_results}{sherlock results table provided by `process_sherlock`}
}
\value{
A table of samples processed with `sample_id`, `assay_id`, `raw_fluorescence`,
`threshold`, and `positive_detection`.
}
\description{
`generate_assay_detection()` generates an assay result table with
positive detections and, depending on the assay, the genetic run type
identification, without interacting with the database
}
\details{
The sherlock results are updated to reflect whether the assays
in a plate run produced raw fluorescence values that exceed the threshold
calculated by `generate_threshold_offline`, resulting in a positive or negative
detection (TRUE or FALSE, where TRUE means the assay is a positive
detection). The function currently uses the positive detection value for the
sample ID at the end of the run time (i.e. the max run time).
}
