% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{generate_final_run_assignment}
\alias{generate_final_run_assignment}
\title{Generate final run assignment table}
\usage{
generate_final_run_assignment(con)
}
\description{
Generate final run assignment table
}
