% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_genetic_detection}
\alias{generate_genetic_detection}
\title{Offline Genetic Identification}
\usage{
generate_genetic_detection(offline_detections_1, offline_detections_2)
}
\arguments{
\item{offline_detections_1}{table of sample IDs, assay IDs, and positive/negative detections to be compared}
}
\description{
`generate_genetic_detection` assigns genetic identifier to a
sample based on assay results.
}
\details{
`generate_genetic_detection` compares two tables of offline
positive detection results, then uses those to assign a genetic identification value.
}
