% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{generate_range}
\alias{generate_range}
\title{Generate a cell range}
\usage{
generate_range(
  table_type = c("raw fluorescence", "background fluorescence", "results"),
  column_header_row
)
}
\arguments{
\item{table_type}{the type of results table, either "raw fluorescence", "background fluorescence", or "results"}

\item{column_header_row}{the row number containing the column headers for the first table of interest}
}
\description{
helper function used to identify a group of cell ranges that encompass
tables containing a type of result from within the Synergy H1 excel workbook output
}
