% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{generate_raw_fl_ranges}
\alias{generate_raw_fl_ranges}
\title{Generate call ranges}
\usage{
generate_raw_fl_ranges(plate_size, wells_used, time_intervals)
}
\arguments{
\item{wells_used}{the number of wells used in the well layout table}
}
\description{
Generate call ranges
}
