% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{generate_subsample_plate_map}
\alias{generate_subsample_plate_map}
\title{Generate Plate Map from Subsample}
\usage{
generate_subsample_plate_map(sample_ids, plate_assay_structure, out_filepath)
}
\arguments{
\item{sample_ids}{list of sample IDs with which to populate a plate map}

\item{plate_assay_structure}{Either \code{dual_assay} or \code{single_assay}}

\item{out_filepath}{Directory to which you want to write the csv(s)}
}
\value{
A named list containing a \code{results} table and a \code{summary} table.
A .csv file with rows \code{A:P} and columns \code{1:24} populated with sample IDs and control blanks
}
\description{
Generates a plate map from a list of sample IDs
}
\details{
Fill this out TODO
}
