% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_threshold}
\alias{generate_threshold}
\title{Generate Thresholds}
\usage{
generate_threshold(
  con,
  plate_run,
  .control_id = "NTC",
  online_mode = TRUE/FALSE,
  offline_sherlock_results = NULL
)
}
\arguments{
\item{con}{valid connection to the database}

\item{plate_run}{plate run object obtained from either `add_plate_run` or `get_plate_run`}

\item{.control_id}{the identifier within the plate run to use as control for calculating thresholds, defaults to "NTC"}

\item{online_mode}{whether the function is being run in online (connected to database) or offline mode}

\item{offline_sherlock_results}{if running offline, this function requires a table of sherlock results}
}
\value{
a table containing thresholds for an event, to be passed to `update_assay_detections()`
}
\description{
`generate_threshold()` calculates the raw fluorescence threshold
values for an assay.
}
\details{
For each assay on a plate run, the threshold value is calculated as two times
the mean value of the last time step from the control blank wells. Each
assay on a plate will have its own control blanks and threshold value.
}
