% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_threshold_offline}
\alias{generate_threshold_offline}
\title{Generate Thresholds Offline}
\usage{
generate_threshold_offline(.control_id = "NTC", offline_sherlock_results)
}
\arguments{
\item{.control_id}{the identifier within the plate run to use as control for calculating thresholds, defaults to "NTC"}

\item{offline_sherlock_results}{if running offline, this function requires a table of sherlock results}
}
\value{
a table containing thresholds for an event, to be passed to `update_assay_detections()`
}
\description{
`generate_threshold_offline()` calculates the raw fluorescence threshold
values for an assay offline
}
\details{
For each assay on a plate run, the threshold value is calculated as two times
the mean value of the last time step from the control blank wells. Each
assay on a plate will have its own control blanks and threshold value.
}
