% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-agency.R
\name{get_agencies}
\alias{get_agencies}
\title{Retrieve Agencies}
\usage{
get_agencies(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_agencies()} returns all agencies within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
agencies <- get_agencies(con)
}
\seealso{
Other agency functions:
\code{\link[=add_agency]{add_agency()}},
\code{\link[=delete_agency]{delete_agency()}},
\code{\link[=update_agency]{update_agency()}},
\code{\link[=update_agency_status]{update_agency_status()}}
}
\concept{agency functions}
