% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplates.R
\name{get_all_subplates_for_run}
\alias{get_all_subplates_for_run}
\title{Get Subplates for Run}
\usage{
get_all_subplates_for_run(
  con,
  plate_run_id,
  .assay_result_table = "assay_result"
)
}
\description{
Get Subplates for Run
}
