% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{get_archive_plates_candidates}
\alias{get_archive_plates_candidates}
\title{Archinve Plates Candidates}
\usage{
get_archive_plates_candidates(con, events, season = get_current_season()$year)
}
\arguments{
\item{con}{a database connection}

\item{season}{season obtained from get_current_season()}
}
\description{
Obtain dataframe of samples ready to be plated onto Archinve Plates
}
