% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample.R
\name{get_clean_dataset}
\alias{get_clean_dataset}
\title{Query database for clean dataset}
\usage{
get_clean_dataset(
  con,
  filtered_samples,
  heterozygote_filter = c(FALSE, TRUE),
  failed_filter = c(FALSE, TRUE)
)
}
\description{
Query database for clean dataset
}
