% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-genetic-method.R
\name{get_genetic_methods}
\alias{get_genetic_methods}
\title{Retrieve Genetic Methods}
\usage{
get_genetic_methods(con, ...)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_genetic_methods()} returns all genetic methods within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
protocols <- get_genetic_methods(con)
}
\seealso{
Other genetic method functions:
\code{\link[=add_genetic_method]{add_genetic_method()}},
\code{\link[=delete_genetic_method]{delete_genetic_method()}},
\code{\link[=update_genetic_method]{update_genetic_method()}},
\code{\link[=update_genetic_method_status]{update_genetic_method_status()}}
}
\concept{genetic method functions}
