% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-genetic-identifaction.R
\name{get_genetic_run_results}
\alias{get_genetic_run_results}
\title{Query Genetic Results}
\usage{
get_genetic_run_results(con, run = NULL, sample_id = NULL, year = NULL)
}
\description{
Get table of results populated in the genetic_run_identification table.
}
