% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{get_hamilton_plates_candidates}
\alias{get_hamilton_plates_candidates}
\title{Hamilton Plate Candidates}
\usage{
get_hamilton_plates_candidates(
  con,
  destination,
  events,
  season = get_current_season()$year
)
}
\description{
Hamilton Plate Candidates
}
