% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.R
\name{get_personnel}
\alias{get_personnel}
\title{Get all personnel}
\usage{
get_personnel(con, collected = TRUE, ...)
}
\arguments{
\item{con}{DBI connection object to database}

\item{collected}{boolean indidcated whether the returned object should be collected from database
or returned as a postgres lazy_query.}

\item{...}{additional named arguments passed into dplyr::filter,
use this add filter to be executed on the database before collecting.}
}
\description{
Get all personnel
}
