% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-protocol.R
\name{get_protocols}
\alias{get_protocols}
\title{Retrieve Protocols}
\usage{
get_protocols(con, ...)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_protocols()} returns all protocols within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
protocols <- get_protocols(con)
}
\seealso{
Other protocol functions:
\code{\link[=add_protocol]{add_protocol()}},
\code{\link[=add_protocol_based_on]{add_protocol_based_on()}},
\code{\link[=delete_protocol]{delete_protocol()}},
\code{\link[=update_protocol]{update_protocol()}},
\code{\link[=update_protocol_status]{update_protocol_status()}}
}
\concept{protocol functions}
