% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-location.R
\name{get_sample_locations}
\alias{get_sample_locations}
\title{Retrieve Sample Locations}
\usage{
get_sample_locations(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_sample_locations()} returns all sample locations within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
sample_locations <- get_sample_locations(con)
}
\seealso{
Other sample location functions:
\code{\link[=add_sample_location]{add_sample_location()}},
\code{\link[=delete_sample_location]{delete_sample_location()}},
\code{\link[=update_sample_location]{update_sample_location()}},
\code{\link[=update_sample_location_status]{update_sample_location_status()}}
}
\concept{sample location functions}
