% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-type.R
\name{get_sample_types}
\alias{get_sample_types}
\title{Retrieve Sample Types}
\usage{
get_sample_types(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_sample_types()} returns all sample types within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
sample_types <- get_sample_types(con)
}
\seealso{
Other sample type functions:
\code{\link[=add_sample_type]{add_sample_type()}},
\code{\link[=delete_sample_type]{delete_sample_type()}},
\code{\link[=update_sample_type]{update_sample_type()}},
\code{\link[=update_sample_type_status]{update_sample_type_status()}}
}
\concept{sample type functions}
