% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{get_spring_winter_samples}
\alias{get_spring_winter_samples}
\title{Samples needing spring/winter assays}
\usage{
get_spring_winter_samples(con)
}
\arguments{
\item{con}{valid connection to database}
}
\value{
A vector of sample IDs.
}
\description{
`get_spring_winter_samples` pulls sample IDs that need the
OTS 16 Spring/Winter assays.
}
\details{
`get_spring_winter_samples` checks the database table `sample_status`
to identify samples that have results for assays 1 and 2 and only had positive
detection for assay 1. These samples require further analysis with the OTS 16 Spring/Winter
assay.
}
