% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample.R
\name{get_unprocessed_dataset}
\alias{get_unprocessed_dataset}
\title{Query database for unprocessed dataset}
\usage{
get_unprocessed_dataset(con, clean_results)
}
\description{
Query database for unprocessed dataset
}
