% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_archive_plate_maps_by_event}
\alias{make_archive_plate_maps_by_event}
\title{Create Plate Maps for Archive Plates}
\usage{
make_archive_plate_maps_by_event(
  con,
  events,
  season = get_current_season()$year,
  output_dir = NULL
)
}
\arguments{
\item{con}{a database connection}

\item{events}{a vector of events to generate plate maps for}

\item{season}{a season to produce plates for, uses current year be default}
}
\description{
Create Plate Maps for Archive Plates
}
