% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_hamilton_plate_maps}
\alias{make_hamilton_plate_maps}
\title{Create Plate Maps for Hamilton Machine}
\usage{
make_hamilton_plate_maps(
  con,
  events,
  destination = c("sherlock", "gtseq"),
  season = get_current_season(),
  output_dir = tempdir()
)
}
\description{
Plate maps will be generated for input into the Hamilton cherry picking machine.
Only samples that are in the "need ots 16" will be used for creating these plate maps
}
