% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_plate_layout}
\alias{make_plate_layout}
\title{Create Plate Layout}
\usage{
make_plate_layout(samples, layout_size = 96)
}
\arguments{
\item{samples}{vector of samples to create map for}

\item{layout_size}{the size of the layout to create}
}
\description{
Create Plate Layout
}
\keyword{internal}
