% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvage-plate-run.R
\name{make_salvage_ids}
\alias{make_salvage_ids}
\title{Create salvage ids}
\usage{
make_salvage_ids(n_samples, year)
}
\arguments{
\item{n_samples}{the total number of samples to create}
}
\description{
Create salvage ids
}
