% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_single_assay_layout}
\alias{make_single_assay_layout}
\title{Make Single Assay Layout}
\usage{
make_single_assay_layout(
  data,
  output_dir,
  season_filter,
  events_name,
  start_index_name_at = 0
)
}
\description{
Make Single Assay Layout
}
