% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_single_assay_layout_for_hamilton}
\alias{make_single_assay_layout_for_hamilton}
\title{Make Single Assay Layout}
\usage{
make_single_assay_layout_for_hamilton(
  data,
  output_dir,
  season_filter,
  events_name
)
}
\description{
Make Single Assay Layout
}
