% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed-genetic-identification.R
\name{ots_early_late_detection}
\alias{ots_early_late_detection}
\title{Determine Run identifcation after Early and Late Assays}
\usage{
ots_early_late_detection(
  con,
  sample_id,
  strategy = c("positive priority", "recent priority")
)
}
\arguments{
\item{con}{a database connection}

\item{sample_id}{the sample id to perform run identification on}

\item{strategy}{the strategy to use when multi same assays are found for a sample see "details" for more}
}
\description{
Determine Run identifcation after Early and Late Assays
}
