% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{parse_plate_flags}
\alias{parse_plate_flags}
\title{Parse Comment for EBK Flags}
\usage{
parse_plate_flags(flag_text, flag_type)
}
\arguments{
\item{flag_text}{string to parse}

\item{flag_type}{what flag to parse for}
}
\description{
Parse plate comment and extract plate related information for EBK id's that fail the qa/qc check
}
