% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{parse_plate_flags_for_EBK_errors}
\alias{parse_plate_flags_for_EBK_errors}
\title{Parse Comment for EBK Flags}
\usage{
parse_plate_flags_for_EBK_errors(text)
}
\arguments{
\item{text}{string to parse}
}
\description{
Parse plate comment and extract plate related information for EBK id's that fail the qa/qc check
}
\keyword{internal}
