% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{partition_df_every_n}
\alias{partition_df_every_n}
\title{Parttion DF every N}
\usage{
partition_df_every_n(df, n)
}
\arguments{
\item{df}{a dataframe}

\item{n}{an integer}
}
\description{
partition a dataframe by every nth row
}
