% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_assay_results}
\alias{process_assay_results}
\title{Process Assay Results}
\usage{
process_assay_results(filepath, range, plate_size, layout)
}
\description{
Process Assay Results
}
