% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherlock.R
\name{process_plate_layout}
\alias{process_plate_layout}
\title{mapping of plate layout location to sample identifier}
\usage{
process_plate_layout(filepath)
}
\description{
mapping of plate layout location to sample identifier
}
