% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_raw_assay_results}
\alias{process_raw_assay_results}
\title{Process Raw Assay Results}
\usage{
process_raw_assay_results(filepath, ranges, plate_size, layout)
}
\description{
Process Raw Assay Results
}
