% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-events.R
\name{process_raw_sample_plan}
\alias{process_raw_sample_plan}
\title{Process Raw Sample Plan}
\usage{
process_raw_sample_plan(filepath)
}
\value{
a list object containing a data frame \code{clean_sample_plan} with the columns
\itemize{
\item location_code
\item sample_event_number
\item first_sample_date
\item sample_bin_code
\item min_fork_length
\item max_fork_length
\item expected_number_of_samples
}

and a data frame \code{location_code_sample_summaries} with the columns
\itemize{
\item location_code
\item max_allowed
\item catch_22
\item catch_23
\item total_samples
}
}
\description{
\code{process_raw_sample_plan}
}
\examples{
# example database connection
process_raw_sample_plan(filepath = "data-raw/2024_raw_sample_plan.xlsx")
}
