% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{process_sherlock}
\alias{process_sherlock}
\title{Process Sherlock Output}
\usage{
process_sherlock(
  sherlock_results_filepath,
  sample_layout_mapping,
  plate_size = c(96, 384)
)
}
\arguments{
\item{filepath}{path to excel file with Sherlock output}
}
\description{
Process Sherlock Output
}
\examples{
sherlock_results_filepath = "data-raw/081022_Chnk_JPE_Early_Plates7-10_results.xlsx"
sample_layout_mapping = readr::read_csv("data-raw/sample_layout_template_384.csv")
plate_size = 384
process_sherlock(sherlock_results_filepath, sample_layout_mapping, plate_size)
}
