% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_sherlock_metadata}
\alias{process_sherlock_metadata}
\title{Process Sherlock Metadata}
\usage{
process_sherlock_metadata(filepath)
}
\description{
helper function for parsing metadata within SynergyH1 output
}
\details{
NOTE: range for metadata may not be static, hopefully we will have a more direct
way to obtain protocol information
}
