% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherquant.R
\name{process_sherquant}
\alias{process_sherquant}
\title{Process SherQuant results}
\usage{
process_sherquant(filepath, sample_type, plate_run_id)
}
\description{
\code{process_sherquant()} processes output from the Sherquant machine
}
