% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_well_sample_details}
\alias{process_well_sample_details}
\title{Get Sample Details}
\usage{
process_well_sample_details(
  filepath,
  sample_type = c("mucus", "fin clip"),
  layout_type = c("split_plate_early_late", "split_plate_late_early",
    "split_plate_spring_winter", "split_plate_winter_spring", "triplicate",
    "single_assay_ots28_early", "single_assay_ots28_late", "single_assay_ots16_spring",
    "single_assay_ots16_winter"),
  plate_run_id,
  assay_order = NULL,
  custom_layout = NULL
)
}
\arguments{
\item{filepath}{Synergy H1 reader output excel file with corresponding plate run layout
in a sheet titled "Plate Map"}

\item{sample_type}{either \code{mucus} or \code{fin_clip}}

\item{layout_type}{either \code{split_plate_early_late}, \code{split_plate_late_early}, \code{split_plate_spring_winter},
\code{split_plate_winter_spring}, \code{triplicate},
\code{single_assay_ots28_early}, \code{single_assay_ots28_late}, \code{single_assay_ots16_spring},
\code{single_assay_ots16_winter}}

\item{plate_run_id}{plate run identifier generated by running \code{add_plate_run()}}
}
\value{
a table to be passed to \code{process_sherlock()}:
\subsection{plate_layout}{
\itemize{
\item location
\item sample_id
\item sample_type_id
\item assay_id
\item plate_run_id
}
}
}
\description{
\code{process_well_sample_details} reads in the plate run layout so sample identifiers
can be mapped to the output of the Synergy H1 reader.
}
\details{
this function is called when the user wants to interact with the database. There is an alternative
version of this function, \code{process_well_sample_details_offline}, that processes sherlock output without
interacting with the database. The output of this function will be used as the \code{sample_details} argument in
\code{process_sherlock()} i.e. it will be used to connect the generic sample identifiers produced
by the Synergy H1 reader to sample identifiers in the database. If the \code{layout_type} argument is
\code{split_plate}, it divide the plate into columns 1:12 and 13:24 and assigns assays based on the
following logic:
\itemize{
\item \code{split_plate_early_late}: columns 1:12 are OTS 28 Early and columns 13:24 are OTS 28 Late
\item \code{split_plate_late_early}: columns 1:12 are OTS 28 Late and columns 13:24 are OTS 28 Early
\item \code{split_plate_spring_winter}: columns 1:12 are OTS 16 Spring and columns 13:24 are OTS 16 Winter
\item \code{split_plate_winter_spring}: columns 1:12 are OTS 16 Winter and columns 13:24 are OTS 16 Spring
}

If the \code{layout_type} argument is \code{triplicate}, the assay ids are assigned as follows:
\itemize{
\item rows A, E, I, and M are OTS 28 Early
\item rows B, F, J, and N are OTS 28 Late
\item rows C, G, K, and O are OTS 16 Spring
\item rows D, H, L, and P are OTS 16 Winter
}

If the \code{layout_type} is one of \code{single_assay_ots28_early}, \code{single_assay_ots28_late}, \code{single_assay_ots16_spring}, or
\code{single_assay_ots16_winter}, all samples processed in that plate run will be assigned that assay.
}
