% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplates.R
\name{remove_subplates_from_run}
\alias{remove_subplates_from_run}
\title{Remove subplate}
\usage{
remove_subplates_from_run(con, plate_run_id, sub_plate)
}
\description{
Given a plate run, delete all of the plate data in database for this subplate
}
