% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch App}
\usage{
run_app(config_path = NULL, ...)
}
\description{
\code{run_app()} launches a shiny application that allows the user to
upload data to the run ID database.
}
\details{
the grunID application can be run when the user has a completed plate run
with assay results. This application calls \code{add_new_plate_results()}.
}
\examples{
grunID::run_app()
}
