% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{run_genetic_identification}
\alias{run_genetic_identification}
\title{Run genetic identification}
\usage{
run_genetic_identification(
  con,
  sample_id = NULL,
  location = NULL,
  year = NULL,
  selection_strategy = "positive priority",
  plate_run_id,
  plate_comment,
  destination_table,
  sample_table,
  results_table,
  sample_status_table,
  layout_type
)
}
\arguments{
\item{con}{a connection to the database}

\item{sample_id}{optional sample id to run identification on}

\item{location}{optional location to identification on}

\item{year}{optional year to identification on}

\item{selection_strategy}{a selection strategy to help resolve cases where many assays exists per sample}
}
\description{
Run genetic identification
}
