% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{run_genetic_identification_gtseq}
\alias{run_genetic_identification_gtseq}
\title{GT-Seq run identification}
\usage{
run_genetic_identification_gtseq(con, samples)
}
\description{
GT-Seq run identification
}
