% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{run_genetic_identification_v3}
\alias{run_genetic_identification_v3}
\title{Run genetic identification with gt seq}
\usage{
run_genetic_identification_v3(con, samples, plate_run_id)
}
\description{
Run genetic identification with gt seq
}
