% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_dataframes}
\alias{template_dataframes}
\alias{protocol_template}
\title{Templates}
\format{
An object of class \code{data.frame} with 1 rows and 17 columns.
}
\usage{
protocol_template
}
\description{
Template data frames to assist users providing the correct
format for adding new items to database
}
\keyword{datasets}
