% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{update_assay_detection}
\alias{update_assay_detection}
\title{Set detection for assay results online}
\usage{
update_assay_detection(con, thresholds, .control_id = "NTC")
}
\arguments{
\item{con}{valid connection to the database}

\item{thresholds}{threshold values calculated in `generate_threshold`}

\item{.control_id}{identifier used to find the control variable}
}
\value{
The number of assay results added to the assay_result table
and the number of samples updated in the genetic_run_identification table.
}
\description{
`update_assay_detection()` updates the assay result table with
positive detections and, depending on the assay, the genetic run type
identification.
}
\details{
The assay result table is updated to reflect whether the assays
in a plate run produced raw fluorescence values that exceed the threshold
calculated by `generate_threshold()`, resulting in a positive or negative
detection (TRUE or FALSE, where TRUE means the assay is a positive
detection). The database is then checked for whether other ots_28 and
ots_16 have been run on the samples. If so, samples are added to
the genetic_run_identification table in the database with a genetic
identification number (see `add_genetic_identification` for details).
}
