% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-protocol.R
\name{update_protocol_status}
\alias{update_protocol_status}
\title{Update Protocol Status}
\usage{
update_protocol_status(con, protocol_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{protocol_id}{A numeric ID for the targeted protocol \code{\link{get_agencies}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_agencies}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_protocol_status()} changes active flag on existing protocol in the protocol lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_agencies <- get_agencies(con)
View(all_agencies) # to view the ID of the protocol needing status change

#deactivate
update_protocol_status(con, 4, set_active=FALSE)
#reactivate
update_protocol_status(con, 4)
}
\seealso{
Other protocol functions:
\code{\link[=add_protocol]{add_protocol()}},
\code{\link[=add_protocol_based_on]{add_protocol_based_on()}},
\code{\link[=delete_protocol]{delete_protocol()}},
\code{\link[=get_protocols]{get_protocols()}},
\code{\link[=update_protocol]{update_protocol()}}
}
\concept{protocol functions}
