% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-run-type.R
\name{update_run_type}
\alias{update_run_type}
\title{Update Run Type}
\usage{
update_run_type(con, run_type_id, run_type)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{run_type_id}{A numeric ID for the targeted run type \code{\link{get_run_types}}}

\item{run_type}{A valid run type dataframe containing the following:
\itemize{
\item \strong{code} Short code representing run type name
\item \strong{run_name} Run type full name
\item \strong{description} Brief description of the run type
}}
}
\description{
\code{update_run_type()} updates an existing run type in the run type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_run_types <- get_run_types(con)
View(all_run_types) # to view the ID of the run type needing updates

updated_run_type <- all_run_types[1, 2:4]
updated_run_type$run_name <- "New Name Run"
update_run_type(con, 1, updated_run_type)
}
\seealso{
Other run type functions:
\code{\link[=add_run_type]{add_run_type()}},
\code{\link[=delete_run_type]{delete_run_type()}},
\code{\link[=get_run_types]{get_run_types()}},
\code{\link[=update_run_type_status]{update_run_type_status()}}
}
\concept{run type functions}
