% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-location.R
\name{update_sample_location}
\alias{update_sample_location}
\title{Update Sample Location}
\usage{
update_sample_location(con, sample_location_id, sample_location)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_location_id}{A numeric ID for the targeted sample location \code{\link{get_sample_locations}}}

\item{sample_location}{A valid sample location dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} A short code for the agency
\item \strong{agency_name} \emph{character} The agency's proper name
}}
}
\description{
\code{update_sample_location()} updates an existing sample location in the sample location lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_locations <- get_sample_locations(con)
View(all_sample_locations) # to view the ID of the sample location needing updates
updated_sample_location <- data.frame(code = "UCD",
                         agency_name = "University of California, Davis")
update_sample_location(con, 4, updated_sample_location)
}
\seealso{
Other sample location functions:
\code{\link[=add_sample_location]{add_sample_location()}},
\code{\link[=delete_sample_location]{delete_sample_location()}},
\code{\link[=get_sample_locations]{get_sample_locations()}},
\code{\link[=update_sample_location_status]{update_sample_location_status()}}
}
\concept{sample location functions}
