% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{update_sample_status}
\alias{update_sample_status}
\title{Update the status code for a list of existing sample}
\usage{
update_sample_status(
  con,
  sample_ids,
  status = c("created", "prepped", "out to field", "return from field", "in analysis",
    "stored", "archived", "other lab")
)
}
\arguments{
\item{con}{connection to the database}

\item{sample_ids}{vector of sample ids to update}

\item{status_code}{status to code to use for update}
}
\description{
Update the status code for a list of existing sample
}
