% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-type.R
\name{update_sample_type}
\alias{update_sample_type}
\title{Update Sample Type}
\usage{
update_sample_type(con, sample_type_id, sample_type)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_type_id}{A numeric ID for the targeted sample type \code{\link{get_sample_types}}}

\item{sample_type}{A valid sample type dataframe containing the following:
\itemize{
\item \strong{code} Short code representing sample type name
\item \strong{sample_type_name} Sample type full name
\item \strong{description} Brief description of the sample type
}}
}
\description{
\code{update_sample_type()} updates an existing sample type in the sample type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_types <- get_sample_types(con)
View(all_sample_types) # to view the ID of the sample type needing updates

updated_sample_type <- all_sample_types[1, 2:4]
updated_sample_type$sample_type_name <- "New Name Run"
update_sample_type(con, 1, updated_sample_type)
}
\seealso{
Other sample type functions:
\code{\link[=add_sample_type]{add_sample_type()}},
\code{\link[=delete_sample_type]{delete_sample_type()}},
\code{\link[=get_sample_types]{get_sample_types()}},
\code{\link[=update_sample_type_status]{update_sample_type_status()}}
}
\concept{sample type functions}
