% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-type.R
\name{update_sample_type_status}
\alias{update_sample_type_status}
\title{Update Sample Type Status}
\usage{
update_sample_type_status(con, sample_type_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_type_id}{A numeric ID for the targeted sample type \code{\link{get_sample_types}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_sample_types}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_sample_type_status()} changes active flag on existing sample type in the sample type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_types <- get_sample_types(con)
View(all_sample_types) # to view the ID of the sample type needing status change

#deactivate
update_sample_type_status(con, 4, set_active=FALSE)
#reactivate
update_sample_type_status(con, 4)
}
\seealso{
Other sample type functions:
\code{\link[=add_sample_type]{add_sample_type()}},
\code{\link[=delete_sample_type]{delete_sample_type()}},
\code{\link[=get_sample_types]{get_sample_types()}},
\code{\link[=update_sample_type]{update_sample_type()}}
}
\concept{sample type functions}
