% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-status-code.R
\name{update_status_code}
\alias{update_status_code}
\title{Update Status Code}
\usage{
update_status_code(con, status_code_id, status_code)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{status_code_id}{A numeric ID for the targeted status code \code{\link{get_status_codes}}}

\item{status_code}{A valid status code dataframe containing the following:
\itemize{
\item \strong{status_code} Status code full name
\item \strong{description} Brief description of the status code
}}
}
\description{
\code{update_status_code()} updates an existing status code in the status code lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_status_codes <- get_status_codes(con)
View(all_status_codes) # to view the ID of the status code needing updates

updated_status_code <- all_status_codes[1, 2:3]
updated_status_code$status_code <- "New all done"
update_status_code(con, 1, updated_status_code)
}
\seealso{
Other status code functions:
\code{\link[=add_status_code]{add_status_code()}},
\code{\link[=delete_status_code]{delete_status_code()}},
\code{\link[=get_sample_status]{get_sample_status()}},
\code{\link[=get_status_codes]{get_status_codes()}},
\code{\link[=set_sample_status]{set_sample_status()}},
\code{\link[=update_status_code_status]{update_status_code_status()}}
}
\concept{status code functions}
