% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-status-code.R
\name{update_status_code_status}
\alias{update_status_code_status}
\title{Update Status Code Status}
\usage{
update_status_code_status(con, status_code_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{status_code_id}{A numeric ID for the targeted status code \code{\link{get_status_codes}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_status_codes}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_status_code_status()} changes active flag on existing status code in the status code lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_status_codes <- get_status_codes(con)
View(all_status_codes) # to view the ID of the status code needing status change

#deactivate
update_status_code_status(con, 4, set_active=FALSE)
#reactivate
update_status_code_status(con, 4)
}
\seealso{
Other status code functions:
\code{\link[=add_status_code]{add_status_code()}},
\code{\link[=delete_status_code]{delete_status_code()}},
\code{\link[=get_sample_status]{get_sample_status()}},
\code{\link[=get_status_codes]{get_status_codes()}},
\code{\link[=set_sample_status]{set_sample_status()}},
\code{\link[=update_status_code]{update_status_code()}}
}
\concept{status code functions}
