% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-run.R
\name{validate_results}
\alias{validate_results}
\title{Validate Results}
\usage{
validate_results(
  con,
  plate_run,
  results_table = c("assay_result", "external_assay_result")
)
}
\arguments{
\item{con}{a connection to the database}

\item{plate_run_id}{plate to run through the validation process}
}
\description{
Validate Results
}
