% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-events.R
\name{validate_sample_bins}
\alias{validate_sample_bins}
\title{sample_bins <- tibble(
                     location_code = c("BTC", "BUT"),
                     sample_event_number = 1:2,
                     first_sample_date = "2020-01-01",
                     sample_bin_code = "A",
                     min_fork_length = 10,
                     max_fork_length = 95,
                     expected_number_of_samples = 10
                     )}
\usage{
validate_sample_bins(con, sample_bins)
}
\description{
sample_bins <- tibble(
                     location_code = c("BTC", "BUT"),
                     sample_event_number = 1:2,
                     first_sample_date = "2020-01-01",
                     sample_bin_code = "A",
                     min_fork_length = 10,
                     max_fork_length = 95,
                     expected_number_of_samples = 10
                     )
}
