% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-genetic-identifaction.R
\name{view_genetic_id_audits}
\alias{view_genetic_id_audits}
\title{View Results Table Audits}
\usage{
view_genetic_id_audits(con, ...)
}
\description{
Display table with data for changes made to the \code{genetic_run_identification} table
}
