% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-assay.R
\name{add_assay}
\alias{add_assay}
\title{Add Assay}
\usage{
add_assay(con, assay)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{assay}{A valid assay dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} A short code for the assay
\item \strong{assay_name} \emph{character} The assay's name
\item \strong{description} \emph{character} Short description of assay
}}
}
\description{
\code{add_assay()} adds a new assay type to the assay lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_assay <- data.frame(code = "OTS28E1",
                        assay_name = "Ots28_Early1",
                        description = "Targets GREB1L region; 58 bp away from
                                       OTS28L1 and targets a SNP"))

add_assay(con, new_assay)
}
\seealso{
Other assay functions: 
\code{\link{delete_assay}()},
\code{\link{get_assays}()},
\code{\link{update_assay_status}()},
\code{\link{update_assay}()}
}
\concept{assay functions}
