% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{add_enum_value}
\alias{add_enum_value}
\title{Add Enumerated Type Value}
\usage{
add_enum_value(
  con,
  enum = c("bin_code_enum", "survey_type_enum", "life_stage_enum", "origin_enum",
    "run_mode_enum", "optics_enum", "light_source_enum", "lamp_energy_enum",
    "well_location_enum"),
  enum_value
)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{enum}{Provide one of the following:
\itemize{
\item \strong{"bin_code_enum"} Used for generating sampling protocols and sample IDs
\item \strong{"survey_type_enum"} Used for describing the survey types in permits
\item \strong{"life_stage_enum"} Used for specifying chinook life stage in permits
\item \strong{"origin_enum"} Used for specifying the chinook origin in permits
\item \strong{"run_mode_enum"} Used for describing the run mode in the Synergy H1 protocol
\item \strong{"optics_enum"} Used for describing the optics setting in the Synergy H1 protocol
\item \strong{"light_source_enum"} Used for describing the light source in the Synergy H1 protocol
\item \strong{"lamp_energy_enum"} Used for describing the lamp energy setting in the Synergy H1 protocol
\item \strong{"well_location_enum"} Used for mapping the Synergy H1 result output to sample
IDs well location in the assay results
}}

\item{enum_value}{new value to be added to existing enum}
}
\description{
\code{add_enum_value()} appends a value to an existing enum
}
\details{
Enumerated (enum) types are data types that comprise a static, ordered set of values.
They are useful for controlling the acceptable values within a column of a database
when a lookup table is overkill.
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

# first confirm the new value is not redundant to existing values
get_enum_values(con, "lamp_energy_enum")

add_enum_value(con, "lamp_energy_enum", "Low")
}
\seealso{
Other enum method functions: 
\code{\link{get_enum_values}()},
\code{\link{update_enum_value}()}
}
\concept{enum method functions}
