% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-run.R
\name{add_new_plate_results}
\alias{add_new_plate_results}
\title{Add New Plate Results}
\usage{
add_new_plate_results(
  con,
  protocol_name,
  genetic_method,
  laboratory,
  lab_work_performed_by,
  description,
  date_run,
  filepath,
  sample_type,
  layout_type,
  plate_size = c(96, 384),
  .control_id = "EBK",
  selection_strategy = "recent priority",
  threshold_strategy = "twice average",
  run_gen_id = FALSE,
  samples_type = c("jpe", "salvage"),
  custom_layout_filepath = NULL
)
}
\arguments{
\item{con}{connection to the database}

\item{genetic_method}{the genetic method used for this run}

\item{laboratory}{the lab used for this run}

\item{description}{a description to ne associated with the plate run}

\item{date_run}{the date that the assat was run}

\item{filepath}{the filepath to the Sherlock results}

\item{sample_type}{the sample type}

\item{layout_type}{the layout that was used for this assay}

\item{plate_size}{either 96 or 384}

\item{protocol}{name pf the protocol used for this run}

\item{is_salvage}{samples are obtained from salvage program}
}
\description{
Add New Plate Results
}
