% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{add_plate_run}
\alias{add_plate_run}
\title{Create Plate Run}
\usage{
add_plate_run(
  con,
  protocol_id,
  genetic_method_id,
  laboratory_id,
  lab_work_performed_by,
  description,
  date_run
)
}
\arguments{
\item{con}{valid connection to the database}

\item{protocol_id}{protocol identifier}

\item{genetic_method_id}{genetic method identifier}

\item{laboratory_id}{laboratory identifier}

\item{lab_work_performed_by}{name of staff who performed the plate run}

\item{description}{a description for the plate run}

\item{date_run}{date of plate run}
}
\description{
`add_plate_run()` adds metadata about a plate run to
the plate_run table in the database.
}
