% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_protocol}
\alias{add_protocol}
\title{Add Protocol}
\usage{
add_protocol(con, protocol)
}
\description{
Add Protocol
}
\seealso{
Other protocol functions: 
\code{\link{delete_protocol}()},
\code{\link{get_protocols}()},
\code{\link{update_protocol}()}
}
\concept{protocol functions}
