% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-location.R
\name{add_sample_location}
\alias{add_sample_location}
\title{Add Sample Location}
\usage{
add_sample_location(con, sample_location)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_location}{A valid sample location dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} A short code for the location
\item \strong{location_name} \emph{character} The location name
\item \strong{stream_name} \emph{character} The stream name (e.g., "American River")
\item \strong{description} \emph{character}
\item \strong{managing_agency_id} \emph{integer} see \code{\link{get_agencies}}
\item \strong{latitude} \emph{numeric}
\item \strong{longitude} \emph{numeric}
}}
}
\description{
\code{add_sample_location()} adds a new sample location to the sample location lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_sample_location <- data.frame(code = "",
                                  location_name = "",
                                  stream_name = "",
                                  description = "",
                                  managing_agency_id = 1,
                                  latitude = 0,
                                  longitude = 0)
add_sample_location(con, new_sample_location)
}
\seealso{
Other sample location functions: 
\code{\link{delete_sample_location}()},
\code{\link{get_sample_locations}()},
\code{\link{update_sample_location_status}()},
\code{\link{update_sample_location}()}
}
\concept{sample location functions}
