% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-events.R
\name{add_sample_plan}
\alias{add_sample_plan}
\title{Add Sample Plan}
\usage{
add_sample_plan(con, sample_plan, verbose = FALSE)
}
\arguments{
\item{con}{A DBI connection object}

\item{sample_plan}{A table containing the following columns from a sample plan:
\code{location_code}, \code{sample_event_number}, \code{first_sample_date}, \code{sample_bin_code}, \code{min_fork_length},
\code{max_fork_length}, and \code{expected_number_of_samples}.}
}
\value{
a named list containing the number of samples added and all unique
sampling event IDs created.
}
\description{
\code{add_sample_plan} registers new sampling events and generates new sample ids
}
\examples{
# example database connection
con <- gr_db_connect()
add_sample_plan(con, sample_plan)
}
