% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-type.R
\name{add_sample_type}
\alias{add_sample_type}
\title{Add Sample Type}
\usage{
add_sample_type(con, sample_type)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{run_type}{A valid sample type dataframe containing the following:
\itemize{
\item \strong{code} Short code representing sample type name
\item \strong{sample_type_name} Sample type full name
\item \strong{description} Brief description of the sample type
}}
}
\description{
\code{add_sample_type()} adds a new sample type to the sample type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

new_sample_type <- data.frame(code = "SFB",
                              sample_type_name = "Some Fish Body",
                              description = "Took a bit of a fish")
add_sample_type(con, new_sample_type)
}
\seealso{
Other sample type functions: 
\code{\link{delete_sample_type}()},
\code{\link{get_sample_types}()},
\code{\link{update_sample_type_status}()},
\code{\link{update_sample_type}()}
}
\concept{sample type functions}
