% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-in-samples.R
\name{check_in_jpe_field_samples}
\alias{check_in_jpe_field_samples}
\title{Check-in Samples from field}
\usage{
check_in_jpe_field_samples(con, filepath, season = year(today()))
}
\description{
Check-in Samples from field
}
