% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-sheet.R
\name{create_field_sheet}
\alias{create_field_sheet}
\title{Create Field Sheets}
\usage{
create_field_sheet(
  wb,
  field_sheet_sample_plan,
  sample_event_number,
  first_sample_date,
  sample_location,
  sample_location_code
)
}
\arguments{
\item{wb}{A Workbook object from \code{\link[openxlsx]{createWorkbook}}}

\item{field_sheet_sample_plan}{A dataframe containing the content of the field sheet,
use the \code{sample_plan} output from \code{\link{get_field_sheet_event_plan}}.
\emph{Note \code{\link{get_field_sheet_event_plan}} also returns the other arguments required by this function.}

The following columns must contain data:
\itemize{
\item \strong{Bin} The sampling bin identifier A-E
\item \strong{Bin FL Range (mm)} The min to max fork length range of the sampling bin
\item \strong{Sample #} A number denoting the order the sample was taken, 1 to the total
number of planned samples with in the sampling bin
\item \strong{Sample ID} \code{{sample_location_code}{YY}_{sample_event_number}_{sample_bin}_{sample_number}} (e.g., "BTC22_3_A_1")
}

The remaining columns ("Date", "Time", "FL (mm)", "Field Run ID", "Fin Clip (Y/N)", "Comments")
will be left empty and filled out by hand in the field}

\item{sample_event_number}{The sampling event number, enumerated from 1 at
the start of each monitoring season for each sampling location}

\item{first_sample_date}{A date object YYYY-MM-DD}

\item{sample_location}{The sampling location name (e.g., "Battle Creek")}

\item{sample_location_code}{The sampling location short code (e.g., "BTC")}
}
\description{
\code{create_field_sheet} appends a worksheet to an existing excel workbook
containing a formatted field sheet to be used by crew collecting genetic samples.
}
\examples{
wb <- openxlsx::createWorkbook()
# Each sample_event will be a tab in a workbook
plan <- get_field_sheet_event_plan(con, sample_event_id = 1)
wb <- create_field_sheet(wb = wb,
                         field_sheet_sample_plan = plan$sample_plan,
                         sample_event_number = plan$sample_event_number,
                         first_sample_date = plan$first_sample_date,
                         sample_location = plan$location_name,
                         sample_location_code = plan$location_code)

plan <- get_field_sheet_event_plan(con, sample_event_id = 2)
wb <- create_field_sheet(wb = wb,
                         field_sheet_sample_plan = plan$sample_plan,
                         sample_event_number = plan$sample_event_number,
                         first_sample_date = plan$first_sample_date,
                         sample_location = plan$location_name,
                         sample_location_code = plan$location_code)

saveWorkbook(wb, "test.xlsx", overwrite = TRUE)
}
