% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agency.R
\name{delete_agency}
\alias{delete_agency}
\title{Delete Agency}
\usage{
delete_agency(con, agency_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{agency_id}{A numeric ID for the targeted agency \code{\link{get_agencies}}}
}
\description{
\code{delete_agency()} deletes an existing agency in the agency lookup table
}
\details{
\strong{Note:} If an assay type has been associated with an assay result record, then
the database restricts deleting this assay type. You must first update those records
with a new assay type before reattempting to delete the assay type. Consider
using the \code{\link{update_assay_status}} function if you are wanting to
retire an assay type while retaining its value for historic records.
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_agencies <- get_agencies(con)
View(all_agencies) # to view the ID of the agency needing deletion

delete_agency(con, 4)
}
\seealso{
Other agency functions: 
\code{\link{add_agency}()},
\code{\link{get_agencies}()},
\code{\link{update_agency_status}()},
\code{\link{update_agency}()}
}
\concept{agency functions}
