% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agency.R
\name{delete_agency}
\alias{delete_agency}
\title{Delete Agency}
\usage{
delete_agency(con, agency_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{agency_id}{A numeric ID for the targeted agency \code{\link{get_agencies}}}
}
\description{
Delete Agency
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_agencies <- get_agencies(con)
View(all_agencies) # to view the ID of the agency needing deletion

delete_agency(con, 4)
}
\seealso{
Other agency functions: 
\code{\link{add_agency}()},
\code{\link{get_agencies}()},
\code{\link{update_agency}()}
}
\concept{agency functions}
