% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-assay.R
\name{delete_assay}
\alias{delete_assay}
\title{Delete Assay}
\usage{
delete_assay(con, assay_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{assay_id}{A numeric ID for the targeted assay \code{\link{get_assays}}}
}
\description{
\code{delete_assay()} deletes an existing assay type in the assay lookup table
}
\details{
\strong{Note:} If an assay type has been associated with an assay result record, then
the database restricts deleting this assay type. You must first update those records
with a new assay type before reattempting to delete the assay type. Consider
using the \code{\link{update_assay_status}} function if you are wanting to
retire an assay type while retaining its value for historic records.
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_assays <- get_assays(con)
View(all_assays) # to view the ID of the assay needing deletion

delete_assay(con, 1)
}
\seealso{
Other assay functions: 
\code{\link{add_assay}()},
\code{\link{get_assays}()},
\code{\link{update_assay_status}()},
\code{\link{update_assay}()}
}
\concept{assay functions}
