% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-genetic-method.R
\name{delete_genetic_method}
\alias{delete_genetic_method}
\title{Delete Genetic Method}
\usage{
delete_genetic_method(con, genetic_method_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{genetic_method_id}{A numeric ID for the targeted genetic method \code{\link{get_genetic_methods}}
\strong{Note:} If an genetic method type has been associated with an assay result record, then
the database restricts deleting this genetic method type. You must first update those records
with a new genetic method type before reattempting to delete the genetic method type. Consider
using the \code{\link{update_genetic_method_status}} function if you are wanting to
retire an genetic method type while retaining its value for historic records.}
}
\description{
\code{delete_genetic_method()} deletes an existing genetic method in the genetic method lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_methods <- get_genetic_methods(con)
View(all_methods) # to view the ID of the genetic method needing updates

delete_genetic_method(con, 1)
}
\seealso{
Other genetic method functions: 
\code{\link{add_genetic_method}()},
\code{\link{get_genetic_methods}()},
\code{\link{update_genetic_method}()},
\code{\link{update_genetic_method_status}()}
}
\concept{genetic method functions}
