% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-laboratory.R
\name{delete_laboratory}
\alias{delete_laboratory}
\title{Delete Laboratory}
\usage{
delete_laboratory(con, laboratory_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{laboratory_id}{A numeric ID for the targeted laboratory \code{\link{get_laboratories}}
\strong{Note:} If an laboratory has been associated with a assay result record, then
the database restricts deleting this laboratory. You must first update those records
with a new laboratory before reattempting to delete the laboratory. Consider
using the \code{\link{update_laboratory_status}} function if you are wanting to
retire an laboratory while retaining its value for historic records.}
}
\description{
\code{delete_laboratory()} deletes an existing laboratory in the laboratory lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_laboratories <- get_laboratories(con)
View(all_laboratories) # to view the ID of the laboratory needing deletion

delete_laboratory(con, 1)
}
\seealso{
Other laboratory functions: 
\code{\link{add_laboratory}()},
\code{\link{get_laboratories}()},
\code{\link{update_laboratory_status}()},
\code{\link{update_laboratory}()}
}
\concept{laboratory functions}
