% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{delete_protocol}
\alias{delete_protocol}
\title{Delete Protocol}
\usage{
delete_protocol(con, protocol_id)
}
\description{
Delete Protocol
}
\seealso{
Other protocol functions: 
\code{\link{add_protocol}()},
\code{\link{get_protocols}()},
\code{\link{update_protocol}()}
}
\concept{protocol functions}
