% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-type.R
\name{delete_sample_type}
\alias{delete_sample_type}
\title{Delete Sample Type}
\usage{
delete_sample_type(con, sample_type_id)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_type_id}{A numeric ID for the targeted sample type \code{\link{get_sample_types}}
\strong{Note:} If an sample type has been associated with a sample, then
the database restricts deleting this sample type. You must first update those records
with a new sample type before reattempting to delete the sample type. Consider
using the \code{\link{update_sample_type_status}} function if you are wanting to
retire an sample type while retaining its value for historic records.}
}
\description{
\code{delete_sample_type()} deletes an existing sample type in the sample type lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_types <- get_sample_types(con)
View(all_sample_types) # to view the ID of the sample type needing deletion

delete_sample_type(con, 1)
}
\seealso{
Other sample type functions: 
\code{\link{add_sample_type}()},
\code{\link{get_sample_types}()},
\code{\link{update_sample_type}()},
\code{\link{update_sample_type_status}()}
}
\concept{sample type functions}
