% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{fill_dual_assay_plate_map}
\alias{fill_dual_assay_plate_map}
\title{fill_dual_assay_plate_map}
\usage{
fill_dual_assay_plate_map(sample_ids)
}
\arguments{
\item{sample_ids}{list of sample IDs with which to populate a plate map}
}
\value{
A list of plate map tables
}
\description{
Generates a plate map from a list of sample IDs
}
\details{
Generate Plate Map for Dual Assay layout

Fill this out TODO
}
