% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{fill_single_assay_plate_map}
\alias{fill_single_assay_plate_map}
\title{fill_single_assay_plate_map}
\usage{
fill_single_assay_plate_map(sample_ids)
}
\arguments{
\item{sample_ids}{list of sample IDs with which to populate a plate map}
}
\value{
A list of plate map tables
}
\description{
Generates a single-assay 384-well plate map from a list of sample IDs
}
\details{
Generate Plate Map for Single Assay v4 layout

This is called in \code{generate_subsample_plate_map()} and fills samples in the 384-well
layout in blocks of four (A1, B1, A2 and B2 would all be the same sample but with different assays).
for more detail on this mapping structure, see \code{grunID::plate_v4_mapping}.
}
