% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{generate_range}
\alias{generate_range}
\title{Generate a cell range}
\usage{
generate_range(
  column_header_row,
  start_col_index,
  start_row_index,
  end_col_index
)
}
\arguments{
\item{wells_used}{the number of wells used in the well layout table}
}
\description{
Generate a cell range
}
