% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{generate_subsample}
\alias{generate_subsample}
\title{Generate Subsample}
\usage{
generate_subsample(con, sampling_event, season)
}
\arguments{
\item{con}{connection to the database}

\item{sampling_event}{A numeric value for the sampling event you are processing (i.e. 1:10)}

\item{season}{The season for which you want to pull subsamples. A season consists of all sampling events
from the given year up to September 30th and from the previous year after October 1st.}
}
\value{
A named list containing a \code{subsample_for_sherlock} table, a \code{subsample_summary} table,
and a \code{remainders_for_gt_seq} table.
The \code{subsample_for_sherlock} table contains all samples selected by the process with the following columns:
\itemize{
\item \strong{sample_id}
\item \strong{datetime_collected}
\item \strong{stream_name}
\item \strong{sample_bin_code}
\item \strong{sample_event_number}
\item \strong{scenario}
}

The \code{subsample_summary} table contains a short summary summarizing the number of samples for a given subsampling scenario by stream,
event, and bin. It contains the following columns:
\itemize{
\item \strong{stream}
\item \strong{event}
\item \strong{bin}
\item \strong{scenario}
\item \strong{subsamples}
}

The \code{remainders_for_gt_seq} table contains all samples \strong{not} selected by the subsampling process with the following columns:
\itemize{
\item \strong{sample_id}
\item \strong{datetime_collected}
\item \strong{stream_name}
\item \strong{sample_bin_code}
\item \strong{sample_event_number}
\item \strong{scenario}
}
}
\description{
Subsamples from a given sample event in a season.
}
\details{
This function takes a subsample from all samples returned from field for a given sampling
event and a given season. It uses the following subsampling logic:
\itemize{
\item At least 50\% of samples per site per event will be sampled.
\item If the number of samples is odd, divide that number by 2 and round the resulting number up to the nearest integer.
\item If the total number of samples for a given site in a given event is less than 20, process all samples for that site/event.
\item If multiple bins are represented in a set of samples for a given site and event, select 50\% of the samples from each bin for processing.
\item If the total number of samples in a bin is less than or equal to 5, process all of the samples for that bin.
\item If this rule contradicts the “less than 20” rule (above), this rule should be prioritized. For example, if we receive a sample set from a given site and event where Bins A, B, C, D, and E are each represented by five samples (total sample size = 25), process all of the samples for that site/event.
\item Subsampling should be random.
}

This function uses \code{set.seed()}, so that the subsample will be the same for a given
\code{sampling_event} and \code{season}.
}
\examples{
# connect with database
con <- gr_db_connect()

# subsample from all sampling events in season 2024
subsamples <- generate_subsample(con, 1, 2024)
}
