% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-subsample.R
\name{generate_subsample}
\alias{generate_subsample}
\title{Set Status Code}
\usage{
generate_subsample(con, sample_event_ids, project)
}
\arguments{
\item{con}{connection to the database}

\item{sample_event_ids}{vector of sample event IDs to subsample}

\item{project}{project (either \code{JPE} or \code{salvage})}
}
\description{
Set the status code for existing samples
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

sample_locations <- set_sample_locations(con,
                                         sample_event_ids = c(9, 10),
                                         project = "JPE")
}
