% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subsample.R
\name{generate_subsample_plate_map}
\alias{generate_subsample_plate_map}
\title{generate_subsample_plate_map}
\usage{
generate_subsample_plate_map(
  sample_ids,
  plate_assay_structure,
  file_basename,
  path
)
}
\arguments{
\item{sample_ids}{list of sample IDs with which to populate a plate map}

\item{plate_assay_structure}{Either \code{dual_assay} or \code{single_assay}}

\item{out_filepath}{Filename. Do not include ".csv" at the end of the filepath.}
}
\value{
A list of filepath(s) to .csv file(s) with rows \code{A:P} and columns \code{1:24} populated with sample IDs and control blanks
}
\description{
Generates a plate map from a list of sample IDs
}
\details{
Generate Plate Map from Subsample

Fill this out TODO
}
