% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_threshold}
\alias{generate_threshold}
\title{Generate Thresholds}
\usage{
generate_threshold(
  con,
  plate_run,
  strategy = "twice average",
  .control_id = "NTC"
)
}
\arguments{
\item{con}{valid connection to the database}

\item{plate_run}{plate run object obtained from either \code{add_plate_run} or \code{get_plate_run}}

\item{strategy}{the strategy to use for computing thresholds, see \strong{details} for more info.}

\item{.control_id}{the identifier within the plate run to use as control for calculating thresholds, defaults to "NTC"}
}
\value{
a table containing thresholds for an event, to be passed to \code{update_assay_detections()}
}
\description{
\code{generate_threshold()} calculates the raw fluorescence threshold
values for an assay.
}
\details{
thresholds can be customized by either selecting from one of the following:
\itemize{
\item "twice average" use twice the average of the control variable as threshold
}

or creating your own threshold function. The function must take a single argument \code{raw_fluorescence} and output a numeric vector.
Example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{custom_threshold <- function(raw_fluorescence) \{
    min(raw_fluorescence)
\}
}\if{html}{\out{</div>}}

This example will use the minimum raw fluorescence value in the plate run as the threshold. When adding to the database the body
of the function will be used to populate the \code{threshold_strategy} column.
}
