% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sherlock.R
\name{generate_threshold}
\alias{generate_threshold}
\title{Generate Thresholds}
\usage{
generate_threshold(con, plate_run_identifier, .control_id = "POS-DNA")
}
\arguments{
\item{con}{valid connection to the database}

\item{plate_run_identifier}{plate run identifier value}
}
\value{
a table containing thresholds for an event, to be passed to `update_assay_detections()`
}
\description{
`generate_threshold()` calculates the raw fluorescence threshold
values for an assay.
}
\details{
For each assay on a plate run, the threshold value is calculated as two times
the mean value of the last time step from the control blank wells. Each
assay on a plate will have its own control blanks and threshold value.
}
