% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-assay.R
\name{get_assays}
\alias{get_assays}
\title{Retrieve Assays}
\usage{
get_assays(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_assays()} returns all assays within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
assays <- get_assays(con)
}
\seealso{
Other assay functions: 
\code{\link{add_assay}()},
\code{\link{delete_assay}()},
\code{\link{update_assay}()},
\code{\link{update_assay_status}()}
}
\concept{assay functions}
