% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-sheet.R
\name{get_field_sheet_event_plan}
\alias{get_field_sheet_event_plan}
\title{Get Sample Event Information for Creating a Field Sheet}
\usage{
get_field_sheet_event_plan(con, sample_event_id)
}
\description{
Get Sample Event Information for Creating a Field Sheet
}
