% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-sheet.R
\name{get_field_sheet_event_plan}
\alias{get_field_sheet_event_plan}
\title{Get Sample Event Information for Creating a Field Sheet}
\usage{
get_field_sheet_event_plan(con, sample_event_id_arg)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_event_id}{The numeric unique identifier of the targeted sampling event for a location
Use \code{\link{get_sample_event}} to query and retrieve the IDs from the database.}
}
\value{
A list object containing the required arguments for \code{\link{create_field_sheet}}
\itemize{
\item \strong{field_sheet_sample_plan} A dataframe containing the content of the field sheet
\item \strong{sample_event_number} The non-unique sample event ID
\item \strong{first_sample_date} A date object YYYY-MM-DD representing the first day of
sampling of a 2 day sampling event
\item \strong{location_name} The sampling location name (e.g., "Battle Creek")
\item \strong{location_code} The sampling location short code (e.g., "BTC")
}
}
\description{
\code{get_field_sheet_event_plan()} retrieves the sampling event information
from the database that is needed to prepare field sheets.
}
\examples{
cfg <- config::get()

con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
plan <- get_field_sheet_event_plan(con, sample_event_id = 1)
}
\seealso{
Other field sheet helpers: 
\code{\link{create_field_sheet}()},
\code{\link{create_multiple_field_sheets}()},
\code{\link{process_field_sheet_samples}()},
\code{\link{update_field_sheet_samples}()}
}
\concept{field sheet helpers}
