% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-run-type.R
\name{get_run_types}
\alias{get_run_types}
\title{Retrieve Run Types}
\usage{
get_run_types(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_run_types()} returns all run types within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
run_types <- get_run_types(con)
}
\seealso{
Other run type functions: 
\code{\link{add_run_type}()},
\code{\link{delete_run_type}()},
\code{\link{update_run_type_status}()},
\code{\link{update_run_type}()}
}
\concept{run type functions}
