% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{get_sample_status}
\alias{get_sample_status}
\title{Get Sample Status}
\usage{
get_sample_status(con, sample_ids = NULL, season = NULL, full_history = FALSE)
}
\arguments{
\item{con}{connection to the database}

\item{sample_ids}{vector of sample ids to update}

\item{full_history}{when set to TRUE, will return the current and each previously assigned status for a sample ID}
}
\description{
View the current or full sample status history
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

sample_locations <- get_sample_locations(con,
                                         sample_ids = c("FTH_RM1722_3_A_1",
                                                         "FTH_RM1722_3_A_2"))
}
\seealso{
Other status code functions: 
\code{\link{add_status_code}()},
\code{\link{delete_status_code}()},
\code{\link{get_status_codes}()},
\code{\link{set_sample_status}()},
\code{\link{update_status_code}()},
\code{\link{update_status_code_status}()}
}
\concept{status code functions}
