% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{get_sample_status}
\alias{get_sample_status}
\title{Get Sample Status}
\usage{
get_sample_status(con, sample_ids, full_history = FALSE)
}
\arguments{
\item{con}{connection to the database}

\item{sample_ids}{vector of sample ids to update}

\item{full_history}{dk}
}
\description{
Get Sample Status
}
