% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample.R
\name{get_samples_by_season}
\alias{get_samples_by_season}
\title{Get Samples by Season}
\usage{
get_samples_by_season(con, season, dataset = c("raw", "clean"))
}
\arguments{
\item{con}{connection to the database}

\item{season}{year in format YYYY. You can pass in a min and max season as c(YYYY, YYYY)}
}
\description{
View sample by season with status and run (if assigned)
}
\examples{
# example database connection
con <- gr_db_connect()
2022_2023_samples <- get_samples_by_season(con, season == c(2022, 2023))
}
