% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sample.R
\name{get_samples_by_season}
\alias{get_samples_by_season}
\title{Get Samples by Season}
\usage{
get_samples_by_season(con, season)
}
\arguments{
\item{con}{connection to the database}

\item{season}{year in format YY}
}
\description{
View sample by season with status and run (if assigned)
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

2022_2023_samples <- get_samples_by_season(con, season == c(2022, 2023))
}
