% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{get_samples_needing_action}
\alias{get_samples_needing_action}
\title{Samples needing further genetic analysis}
\usage{
get_samples_needing_action(con)
}
\arguments{
\item{con}{valid connection to database}
}
\value{
A list with the following elements, each containing a list of Sample IDs in
that category:
\itemize{
\item needs_ots_16
\item repeat_ots_28
\item repeat_ots_16
\item potential_heterozygotes
}
}
\description{
\code{get_samples_needing_action} pulls sample IDs that need further
genetic analysis.
}
\details{
\code{get_samples_needing_action} checks the database table \code{sample_status}
to identify samples in the following categories:
\itemize{
\item Have been run through OTS 28 Early/Late and need to be run through OTS 16 Spring/Winter
\item Have been run through OTS 28 Early/Late and produced a negative result for both, indicating
they require to be run again
\item Have been run through OTS 16 Spring/Winter and produced a negative result for both,
indicating they require to be run again
\item Produced positive results for both assays either within OTS 28 Early/Late OR
OTS 16 Spring/Winter, indicating they are potential heterozygotes
}
}
