% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-status-code.R
\name{get_status_codes}
\alias{get_status_codes}
\title{Retrieve Status Codes}
\usage{
get_status_codes(con, is_active = TRUE, all_results = FALSE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\description{
\code{get_status_codes()} returns all status codes within the database
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
sample_types <- get_status_codes(con)
}
\seealso{
Other status code functions: 
\code{\link{add_status_code}()},
\code{\link{delete_status_code}()},
\code{\link{get_sample_status}()},
\code{\link{set_sample_status}()},
\code{\link{update_status_code_status}()},
\code{\link{update_status_code}()}
}
\concept{status code functions}
