% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{gr_db_connect}
\alias{gr_db_connect}
\title{Connect to Run ID Database}
\usage{
gr_db_connect(
  config = db_get_config(),
  username = NULL,
  password = NULL,
  host = NULL,
  port = NULL
)
}
\description{
Connect to Run ID Database
}
