% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-sheet.R
\name{is_valid_sample_field_data}
\alias{is_valid_sample_field_data}
\title{Helper function - checks field sheet data format}
\usage{
is_valid_sample_field_data(data)
}
\description{

}
\details{
Checks whether the connection is valid using \code{\link{is_valid_con}} and
checks class of input variables.
}
