% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-maps.R
\name{make_dual_assay_layout}
\alias{make_dual_assay_layout}
\title{Make Dual Assay Layout}
\usage{
make_dual_assay_layout(
  data,
  layout_size = 96,
  output_dir,
  season_filter,
  events_name,
  plate_name_offset = 0
)
}
\description{
Make Dual Assay Layout
}
