% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-identification.R
\name{ots_early_late_detection}
\alias{ots_early_late_detection}
\title{Determine Run identifcation after Early and Late Assays}
\usage{
ots_early_late_detection(
  con,
  sample_id,
  selection_strategy = c("positive priority", "recent priority")
)
}
\arguments{
\item{con}{a database connection}

\item{sample_id}{the sample id to perform run identification on}

\item{strategy}{the strategy to use when multi same assays are found for a sample see "details" for more}
}
\description{
Determine Run identifcation after Early and Late Assays
}
\details{
The system allows users to submit as many assays per sample as desired, however at the end in order to assign
a genetic run type to a sample a single early, late, spring, and winter assay must be used. The selection
strategy gives the function a way to select when it finds multiple assays for a sample.
\itemize{
\item "positive priority" - will choose the the positive results from the available assays needed
\item "recent priority" - will choose the most recent assays from the available assays needed.
}
}
