% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_plate_layout}
\alias{process_plate_layout}
\title{Process Plate Layout}
\usage{
process_plate_layout(filepath, plate_size)
}
\description{
helper function for mapping of plate layout location to sample identifier
}
