% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{process_sherlock}
\alias{process_sherlock}
\title{Process Sherlock Output}
\usage{
process_sherlock(
  sherlock_results_filepath,
  sample_layout_mapping,
  plate_size = c(96, 384)
)
}
\arguments{
\item{filepath}{path to excel file with Sherlock output}
}
\description{
Process Sherlock Output
}
