% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_sherlock}
\alias{process_sherlock}
\title{Process Sherlock Results}
\usage{
process_sherlock(filepath, sample_details, plate_size = c(96, 384))
}
\arguments{
\item{filepath}{path to excel file generated by Synergy H1 software}

\item{sample_details}{dataframe containing sample information at each well,
see the \strong{Sample Details} section for more information}

\item{plate_size}{either 96 or 384}
}
\value{
A table to be passed to \code{add_raw_assay_results()}:
\subsection{1) raw_assay_results}{
\itemize{
\item sample_id
\item raw_fluorescence
\item background_value
\item time
\item plate_run_id
\item well_location
}
}
}
\description{
\code{process_sherlock()} extracts and transforms result data from the Synergy H1 reader
}
\section{Sample Details}{
The Synergy H1 result output from plate runs is organized using generic identifiers
that need to be mapped to the real sample identifiers in order to properly associate
the results with the correct sample id in the database.The output file’s
“Layout” section contains the well locations for samples encoded as non-unique generic
sample ids (e.g., SPL1, BLK). The number of rows and columns in the Layout vary
based on the plate size. A plate size 96 has rows A:H and columns 1:12 and
a plate size 384 has rows A:P and columns 1:24. The results for each sample are
found within columns that are the letter-number combination of the well location
captured in the Layout section (e.g., if SPL1 is in row A and column 3,
then the measurements at each time step are recorded in a column labeled A3).
To map the sample ids to the generic identifiers produced by the Synergy H1 software
(e.g., SPL1 and A3), the user must provide the true sample ids and other metadata
at each well location in the as a dataframe containing these columns:
\itemize{
\item \strong{location:} \emph{(character)} letter-number combination based on Synergy H1 Layout grid (e.g., A3)
\item \strong{sample_id:} \emph{(character)} unique ID of mucus swab or fin clip made up of site code, year,
sample event, sampling bin, and count within sampling bin (e.g., BTC22_3_A_1)
\item \strong{sample_type:} \emph{(numeric)} provide 1 for mucus and 2 for fin clip
\item \strong{assay_id:} \emph{(numeric)} the ID for assay types (e.g., 1 = OTS28 Early 1),
use \code{get_assay()} to view all available options.
\item \strong{plate_run_id:} \emph{(numeric)} the ID returned from running \code{add_plate_run()},
the function that creates a new plate run record in the database to associate
important metadata with the assay results
}
}

\examples{
sample_details = readr::read_csv("data-raw/sample_layout_template_384.csv")
process_sherlock(filepath = "data-raw/081022_Chnk_JPE_Early_Plates7-10_results.xlsx",
                 sample_details = sample_details,
                 plate_size = 384)
}
