% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-sherlock.R
\name{process_sherlock}
\alias{process_sherlock}
\title{Process Sherlock Results}
\usage{
process_sherlock(filepath, sample_layout_mapping, plate_size = c(96, 384))
}
\arguments{
\item{filepath}{path to excel file generated by Synergy H1 software}

\item{sample_layout_mapping}{dataframe containing the mapping of well locations and sample identifiers}

\item{plate_size}{either 96 or 384}
}
\value{
A list containing the following 3 tables:
\subsection{1) metadata}{
\itemize{
\item sample_id
\item raw_fluorescence
\item background_value
\item time
\item plate_run_id
\item well_location
}
}

\subsection{2) raw_assay_results}{
\itemize{
\item sample_id
\item raw_fluorescence
\item background_value
\item time
\item plate_run_id
\item well_location
}
}

\subsection{3) assay_results}{
\itemize{
\item sample_id
\item sample_type_id
\item assay_id
\item rfu_back_subtracted
\item plate_run_id
\item well_location
}
}
}
\description{
\code{process_sherlock()} extracts and transforms result data from the Synergy H1 reader
}
\section{Sample Layout Mapping}{
Plate Size 96 rows A:H and columns 1:12
Plate Size 384 rows A:P and columns 1:24
\itemize{
\item location
\item sample_id
\item sample_type_id
\item assay_id
\item plate_run_id
}
}

\examples{
sample_layout_mapping = readr::read_csv("data-raw/sample_layout_template_384.csv")
process_sherlock(filepath = "data-raw/081022_Chnk_JPE_Early_Plates7-10_results.xlsx",
                 sample_layout_mapping = sample_layout_mapping,
                 plate_size = 384)
}
