% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_sherlock}
\alias{process_sherlock}
\title{Process Sherlock Results}
\usage{
process_sherlock(filepath, sample_layout_mapping, plate_size = c(96, 384))
}
\arguments{
\item{filepath}{path to excel file generated by Synergy H1 software}

\item{sample_layout_mapping}{dataframe containing the mapping of well locations and sample identifiers}

\item{plate_size}{either 96 or 384}
}
\value{
A list containing the following 3 tables:
\subsection{1) metadata}{
\itemize{
\item sample_id
\item raw_fluorescence
\item background_value
\item time
\item plate_run_id
\item well_location
}
}

\subsection{2) raw_assay_results}{
\itemize{
\item sample_id
\item raw_fluorescence
\item background_value
\item time
\item plate_run_id
\item well_location
}
}

\subsection{3) assay_results}{
\itemize{
\item sample_id
\item sample_type_id
\item assay_id
\item rfu_back_subtracted
\item plate_run_id
\item well_location
}
}
}
\description{
\code{process_sherlock()} extracts and transforms result data from the Synergy H1 reader
}
\section{Sample Layout Mapping}{
The Synergy H1 result output from plate runs is organized using generic identifiers
that need to be mapped to the real sample identifiers in order to properly associate
the results with the correct sample id in the database.The output file’s
“Layout” section contains the well locations for samples encoded as non-unique generic
sample ids (e.g., SPL1, BLK). The number of rows and columns in the Layout vary
based on the plate size. A plate size 96 has rows A:H and columns 1:12 and
a plate size 384 has rows A:P and columns 1:24. The results for each sample are
found within columns that are the letter-number combination of the well location
captured in the Layout section (e.g., if SPL1 is in row A and column 3,
then the measurements at each time step are recorded in a column labeled A3).
To map the sample ids to the generic identifiers produced by the Synergy H1 software
(e.g., SPL1 and A3), the user must provide the true sample ids at each well
location in the following format:
\itemize{
\item location -
\item sample_id -
\item sample_type -
\item assay_id -
\item plate_run_id -
}
}

\examples{
sample_layout_mapping = readr::read_csv("data-raw/sample_layout_template_384.csv")
process_sherlock(filepath = "data-raw/081022_Chnk_JPE_Early_Plates7-10_results.xlsx",
                 sample_layout_mapping = sample_layout_mapping,
                 plate_size = 384)
}
