% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transform-sherlock.R
\name{process_well_sample_details}
\alias{process_well_sample_details}
\title{Get Sample Details}
\usage{
process_well_sample_details(filepath, sample_type, assay_type, plate_run_id)
}
\arguments{
\item{filepath}{Synergy H1 reader output excel file with corresponding plate run layout
in a sheet titled "Plate Map"}

\item{sample_type}{either \code{mucus} or \code{fin_clip}}

\item{assay_type}{either \code{Ots28_Early1}, \code{Ots28_Late1}, \code{Ots16_Spring1}, or \code{Ots16_Winter1}.
This is set here.}

\item{plate_run_id}{plate run identifier generated by running \code{add_plate_run()}.}
}
\value{
a table to be passed to \code{process_sherlock()}:
\subsection{plate_layout}{
\itemize{
\item location
\item sample_id
\item sample_type_id
\item assay_id
\item plate_run_id
}
}
}
\description{
\code{process_well_sample_details} reads in the plate run layout so sample identifiers
can be mapped to the output of the Synergy H1 reader.
}
\details{
the output of this function will be used as the \code{sample_details} argument in
\code{process_sherlock()} i.e. it will be used to connect the generic sample identifiers produced
by the Synergy H1 reader to sample identifiers in the database.
}
