% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{set_sample_status}
\alias{set_sample_status}
\title{Set the status code for a list of existing samples}
\usage{
set_sample_status(
  con,
  sample_ids,
  status = c("created", "prepped", "out to field", "return from field", "in analysis",
    "stored", "archived", "other lab"),
  comment = NULL
)
}
\arguments{
\item{con}{connection to the database}

\item{sample_ids}{vector of sample ids to update}

\item{status}{status to code to use for update}
}
\description{
Set the status code for a list of existing samples
}
