% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-status.R
\name{set_sample_status}
\alias{set_sample_status}
\title{Set Status Code}
\usage{
set_sample_status(con, sample_ids, status_code_id, comment = NULL)
}
\arguments{
\item{con}{connection to the database}

\item{sample_ids}{vector of sample ids to update}

\item{status}{status code to use for the update
\itemize{
\item \strong{created} status assigned automatically when a new sample ID is generated
\item \strong{prepped} sample tube has been created
\item \strong{out to field} sample tube is out in the field to be collected
\item \strong{return from field} sample tube has returned from the field
\item \strong{in analysis} sample is being analyzed
\item \strong{stored} sample is in storage
\item \strong{archived} sample has been sent to a tissue archive
\item \strong{other lab} sample is out at another lab
}}
}
\description{
Set the status code for existing samples
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

sample_locations <- set_sample_locations(con,
                                         sample_ids = c("FTH_RM1722_3_A_1",
                                                         "FTH_RM1722_3_A_2"),
                                         status = "prepped",
                                         comment = "ready for pickup")
}
\seealso{
Other status code functions: 
\code{\link{add_status_code}()},
\code{\link{delete_status_code}()},
\code{\link{get_sample_status}()},
\code{\link{get_status_codes}()},
\code{\link{update_status_code}()},
\code{\link{update_status_code_status}()}
}
\concept{status code functions}
