% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agency.R
\name{update_agency}
\alias{update_agency}
\title{Update Agency}
\usage{
update_agency(con, agency_id, agency)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{agency_id}{A numeric ID for the targeted agency \code{\link{get_agencies}}}

\item{agency}{A valid agency dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} A short code for the agency
\item \strong{agency_name} \emph{character} The agency's proper name
}}
}
\description{
\code{update_agency()} updates an existing agency in the agency lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_agencies <- get_agencies(con)
View(all_agencies) # to view the ID of the agency needing updates
updated_agency <- data.frame(code = "UCD",
                         agency_name = "University of California, Davis"))
update_agency(con, 4, updated_agency)
}
\seealso{
Other agency functions: 
\code{\link{add_agency}()},
\code{\link{delete_agency}()},
\code{\link{get_agencies}()},
\code{\link{update_agency_status}()}
}
\concept{agency functions}
