% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-assay.R
\name{update_assay}
\alias{update_assay}
\title{Update Assay}
\usage{
update_assay(con, assay_id, assay)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{assay_id}{A numeric ID for the targeted assay \code{\link{get_assays}}}

\item{assay}{A valid assay dataframe with the following:
\enumerate{
\item \strong{code} \emph{character} A short code for the assay
\item \strong{assay_name} \emph{character} The assay's name
\item \strong{description} \emph{character} Short description of assay
}}
}
\description{
\code{update_assay()} updates an existing assay type in the assay lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_assays <- get_assays(con)
View(all_assays) # to view the ID of the assay needing updates
updated_assay <- data.frame(code = "OTS28E1",
                            assay_name = "Ots28 Early 1",
                            description = "Targets GREB1L region"))

update_assay(con, 1, updated_assay)
}
\seealso{
Other assay functions: 
\code{\link{add_assay}()},
\code{\link{delete_assay}()},
\code{\link{get_assays}()},
\code{\link{update_assay_status}()}
}
\concept{assay functions}
