% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-assay.R
\name{update_assay_status}
\alias{update_assay_status}
\title{Update Assay Status}
\usage{
update_assay_status(con, assay_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{assay_id}{A numeric ID for the targeted assay \code{\link{get_assays}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_assays}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_assay_status()} changes active flag on existing assay type in the assay lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_assays <- get_assays(con)
View(all_assays) # to view the ID of the assay needing status change

#deactivate
update_assay_status(con, 1, set_active=FALSE)
#reactivate
update_assay_status(con, 1)
}
\seealso{
Other assay functions: 
\code{\link{add_assay}()},
\code{\link{delete_assay}()},
\code{\link{get_assays}()},
\code{\link{update_assay}()}
}
\concept{assay functions}
