% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-sheet.R
\name{update_field_sheet_samples}
\alias{update_field_sheet_samples}
\title{Update Sample Field Sheet Data}
\usage{
update_field_sheet_samples(con, field_data)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}
}
\value{
Does not return any objects.
}
\description{
\code{update_field_sheet_samples()} takes a formatted tibble of field sample
data and updates those samples in the database.
}
\details{
This function requires a valid connection to the database and a processed dataset
with all variables in the correct format. Errors may be due to invalid data structure or
connection; see \code{\link{is_valid_con}} and \code{\link{is_valid_sample_field_data}} for more
information. Variables should be as follows:
\itemize{
\item \strong{sample_id} The unique identifier for the sample recorded, of class "character".
\item \strong{datetime_collected} The date and time (YYYY-MM-DD H:M:S) the sample was processed,
of class "datetime" or ("POSIXct" "POSIXt")
\item \strong{fork_length_mm} The recorded fork length corresponding to the sample, of class "numeric".
\item \strong{field_run_type_id} The unique identifier for the field run type, of class "numeric".
\item \strong{fin_clip} Logical variable indicating whether a fin clip sample was taken, of class "logical".
\item \strong{field_comment} Any recorded comments from the field regarding the sample, of class "character".
}
}
\examples{
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

filepath <- "data-raw/test.xlsx"
field_data_clean <- process_field_sheet_samples(filepath)
update_field_sheet_samples(con, field_data_clean)
}
\seealso{
Other field sheet helpers: 
\code{\link{create_field_sheet}()},
\code{\link{create_multiple_field_sheets}()},
\code{\link{get_field_sheet_event_plan}()},
\code{\link{process_field_sheet_samples}()}
}
\concept{field sheet helpers}
