% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-method.R
\name{update_genetic_method_status}
\alias{update_genetic_method_status}
\title{Change Agency Status}
\usage{
update_genetic_method_status(con, genetic_method_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{genetic_method_id}{A numeric ID for the targeted genetic method \code{\link{get_agencies}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_agencies}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_genetic_method_status()} changes active flag on existing genetic method in the genetic method lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_agencies <- get_agencies(con)
View(all_agencies) # to view the ID of the genetic method needing status change

#deactivate
update_genetic_method_status(con, 4, set_active=FALSE)
#reactivate
update_genetic_method_status(con, 4)
}
\seealso{
Other genetic method functions: 
\code{\link{add_genetic_method}()},
\code{\link{delete_genetic_method}()},
\code{\link{get_genetic_methods}()},
\code{\link{update_genetic_method}()}
}
\concept{genetic method functions}
