% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-genetic-identifaction.R
\name{update_genetic_run_id}
\alias{update_genetic_run_id}
\title{Update Genetic Run ID}
\usage{
update_genetic_run_id(con, sample_id, run_type)
}
\description{
Updates genetic id for a sample in the database
}
