% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-laboratory.R
\name{update_laboratory}
\alias{update_laboratory}
\title{Update Laboratory}
\usage{
update_laboratory(con, laboratory_id, laboratory)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{laboratory_id}{A numeric ID for the targeted laboratory \code{\link{get_laboratories}}}

\item{laboratory}{A valid laboratory dataframe containing the following:
\itemize{
\item \strong{code} Short code representing laboratory name
\item \strong{laboratory_name} Laboratory full name
\item \strong{description} Brief description of the laboratory
}}
}
\description{
\code{update_laboratory()} updates an existing laboratory in the laboratory lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_laboratories <- get_laboratories(con)
View(all_laboratories) # to view the ID of the laboratory needing updates

updated_laboratory <- all_laboratories[1, 2:4]
updated_laboratory$laboratory_name <- "The Department of Water Resources Genetic Monitoring Laboratory"
update_laboratory(con, 1, updated_laboratory)
}
\seealso{
Other laboratory functions: 
\code{\link{add_laboratory}()},
\code{\link{delete_laboratory}()},
\code{\link{get_laboratories}()},
\code{\link{update_laboratory_status}()}
}
\concept{laboratory functions}
