% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{update_protocol}
\alias{update_protocol}
\title{Update Protocol}
\usage{
update_protocol(con, protocol_id, protocol)
}
\description{
Update Protocol
}
\seealso{
Other protocol functions: 
\code{\link{add_protocol}()},
\code{\link{delete_protocol}()},
\code{\link{get_protocols}()}
}
\concept{protocol functions}
