% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-protocol.R
\name{update_protocol}
\alias{update_protocol}
\title{Update Protocol}
\usage{
update_protocol(con, protocol_id, protocol)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{protocol_id}{A numeric ID for the targeted protocol \code{\link{get_protocols}}}

\item{protocol}{A valid protocol dataframe, reference \code{\link{protocol_template}} as an example}
}
\description{
\code{update_protocol()} updates an existing protocol in the protocol lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_protocols <- get_protocols(con)
View(all_protocols) # to view the ID of the protocol needing updates

updated_protocol <- protocol_template
updated_protocol$software_version <- "3.11.20"
update_protocol(con, 1, updated_protocol)
}
\seealso{
Other protocol functions: 
\code{\link{add_protocol}()},
\code{\link{add_protocol_based_on}()},
\code{\link{delete_protocol}()},
\code{\link{get_protocols}()},
\code{\link{update_protocol_status}()}
}
\concept{protocol functions}
