% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu-sample-location.R
\name{update_sample_location_status}
\alias{update_sample_location_status}
\title{Update Sample Location Status}
\usage{
update_sample_location_status(con, sample_location_id, set_active = TRUE)
}
\arguments{
\item{con}{A DBI connection object obtained from DBI::dbConnect()}

\item{sample_location_id}{A numeric ID for the targeted sample location \code{\link{get_sample_locations}}}

\item{set_active}{A boolean, TRUE for activating and FALSE for deactivating.
When a record is active, it is returned by default when \code{\link{get_sample_locations}}
is called. This helps preserve look up values that are valid in historic
contexts, but are no longer valid for current data records.}
}
\description{
\code{update_sample_location_status()} changes active flag on existing sample location in the sample location lookup table
}
\examples{
# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_locations <- get_sample_locations(con)
View(all_sample_locations) # to view the ID of the sample location needing status change

#deactivate
update_sample_location_status(con, 4, set_active=FALSE)
#reactivate
update_sample_location_status(con, 4)
}
\seealso{
Other sample location functions: 
\code{\link{add_sample_location}()},
\code{\link{delete_sample_location}()},
\code{\link{get_sample_locations}()},
\code{\link{update_sample_location}()}
}
\concept{sample location functions}
